# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech E-SERVICES SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  balance: number;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface AddBalanceBody {
  userId: number;
  amount: number;
}

export type ProviderStatus =
  (typeof ProviderStatus)[keyof typeof ProviderStatus];

export const ProviderStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Provider {
  id: number;
  name: string;
  apiUrl: string;
  apiKey: string;
  status: ProviderStatus;
  createdAt: string;
}

export type CreateProviderBodyStatus =
  (typeof CreateProviderBodyStatus)[keyof typeof CreateProviderBodyStatus];

export const CreateProviderBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateProviderBody {
  name: string;
  apiUrl: string;
  apiKey: string;
  status?: CreateProviderBodyStatus;
}

export type UpdateProviderBodyStatus =
  (typeof UpdateProviderBodyStatus)[keyof typeof UpdateProviderBodyStatus];

export const UpdateProviderBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateProviderBody {
  name?: string;
  apiUrl?: string;
  apiKey?: string;
  status?: UpdateProviderBodyStatus;
}

export interface SyncResult {
  added: number;
  updated: number;
  deactivated: number;
  errors: string[];
}

export type ServiceStatus = (typeof ServiceStatus)[keyof typeof ServiceStatus];

export const ServiceStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Service {
  id: number;
  providerServiceId: string;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  name: string;
  category: string;
  type: string;
  originalRate: number;
  sellRate: number;
  min: number;
  max: number;
  status: ServiceStatus;
  createdAt: string;
}

export type UpdateServiceBodyStatus =
  (typeof UpdateServiceBodyStatus)[keyof typeof UpdateServiceBodyStatus];

export const UpdateServiceBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateServiceBody {
  name?: string;
  category?: string;
  sellRate?: number;
  status?: UpdateServiceBodyStatus;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  partial: "partial",
  canceled: "canceled",
} as const;

export interface Order {
  id: number;
  userId: number;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  /** @nullable */
  providerOrderId?: string | null;
  link: string;
  quantity: number;
  charge: number;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateOrderBody {
  serviceId: number;
  link: string;
  quantity: number;
}

export type ListServicesParams = {
  category?: string;
};