# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech E-SERVICES SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    balance: zod.number(),
    role: zod.enum(["user", "admin"]),
    createdAt: zod.date(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  balance: zod.number(),
  role: zod.enum(["user", "admin"]),
  createdAt: zod.date(),
});

/**
 * @summary Add balance to user (Admin only)
 */
export const AddBalanceBody = zod.object({
  userId: zod.number(),
  amount: zod.number(),
});

export const AddBalanceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  balance: zod.number(),
  role: zod.enum(["user", "admin"]),
  createdAt: zod.date(),
});

/**
 * @summary List all users (Admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  balance: zod.number(),
  role: zod.enum(["user", "admin"]),
  createdAt: zod.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary List all providers (Admin only)
 */
export const ListProvidersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  apiUrl: zod.string(),
  apiKey: zod.string(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.date(),
});
export const ListProvidersResponse = zod.array(ListProvidersResponseItem);

/**
 * @summary Create a new provider (Admin only)
 */
export const createProviderBodyStatusDefault = `active`;

export const CreateProviderBody = zod.object({
  name: zod.string(),
  apiUrl: zod.string(),
  apiKey: zod.string(),
  status: zod
    .enum(["active", "inactive"])
    .default(createProviderBodyStatusDefault),
});

/**
 * @summary Update a provider (Admin only)
 */
export const UpdateProviderParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProviderBody = zod.object({
  name: zod.string().optional(),
  apiUrl: zod.string().optional(),
  apiKey: zod.string().optional(),
  status: zod.enum(["active", "inactive"]).optional(),
});

export const UpdateProviderResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  apiUrl: zod.string(),
  apiKey: zod.string(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.date(),
});

/**
 * @summary Delete a provider (Admin only)
 */
export const DeleteProviderParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Trigger service sync from all providers (Admin only)
 */
export const SyncServicesResponse = zod.object({
  added: zod.number(),
  updated: zod.number(),
  deactivated: zod.number(),
  errors: zod.array(zod.string()),
});

/**
 * @summary List all active services
 */
export const ListServicesQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  providerServiceId: zod.string(),
  providerId: zod.number(),
  providerName: zod.string().nullish(),
  name: zod.string(),
  category: zod.string(),
  type: zod.string(),
  originalRate: zod.number(),
  sellRate: zod.number(),
  min: zod.number(),
  max: zod.number(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.date(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary List unique service categories
 */
export const ListCategoriesResponseItem = zod.string();
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get a single service
 */
export const GetServiceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetServiceResponse = zod.object({
  id: zod.number(),
  providerServiceId: zod.string(),
  providerId: zod.number(),
  providerName: zod.string().nullish(),
  name: zod.string(),
  category: zod.string(),
  type: zod.string(),
  originalRate: zod.number(),
  sellRate: zod.number(),
  min: zod.number(),
  max: zod.number(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.date(),
});

/**
 * @summary Update a service (Admin only)
 */
export const UpdateServiceParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateServiceBody = zod.object({
  name: zod.string().optional(),
  category: zod.string().optional(),
  sellRate: zod.number().optional(),
  status: zod.enum(["active", "inactive"]).optional(),
});

export const UpdateServiceResponse = zod.object({
  id: zod.number(),
  providerServiceId: zod.string(),
  providerId: zod.number(),
  providerName: zod.string().nullish(),
  name: zod.string(),
  category: zod.string(),
  type: zod.string(),
  originalRate: zod.number(),
  sellRate: zod.number(),
  min: zod.number(),
  max: zod.number(),
  status: zod.enum(["active", "inactive"]),
  createdAt: zod.date(),
});

/**
 * @summary List orders (user sees own, admin sees all)
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  serviceId: zod.number(),
  serviceName: zod.string().nullish(),
  providerOrderId: zod.string().nullish(),
  link: zod.string(),
  quantity: zod.number(),
  charge: zod.number(),
  status: zod.enum([
    "pending",
    "in_progress",
    "completed",
    "partial",
    "canceled",
  ]),
  createdAt: zod.date(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Place a new order
 */
export const CreateOrderBody = zod.object({
  serviceId: zod.number(),
  link: zod.string(),
  quantity: zod.number(),
});

/**
 * @summary Get a single order
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  serviceId: zod.number(),
  serviceName: zod.string().nullish(),
  providerOrderId: zod.string().nullish(),
  link: zod.string(),
  quantity: zod.number(),
  charge: zod.number(),
  status: zod.enum([
    "pending",
    "in_progress",
    "completed",
    "partial",
    "canceled",
  ]),
  createdAt: zod.date(),
});